Fix user feedback: DX, strict_types, silent errors, sentinel:test
- Fix strict_types TypeError: cast (int)/(bool) on all config values
  assigned to typed properties (AlertDeduplicator, RateLimiter,
  StatsdCollector, MonitoringService, ContextSanitizer)
- Make Slack/Discord log channel level configurable via config
  (sentinel.slack.level / sentinel.discord.level), default to 'debug'
  so sentinel:test works out of the box
- Add 'sentinel' container alias so app('sentinel') works
- Replace silent catches with Log::warning in sendToChannels so
  webhook failures are visible in Laravel logs
- Throw RuntimeException in non-production when DB storage fails
  (missing migration), keep silent in production
- Rewrite sentinel:test: direct HTTP webhook connectivity test before
  sending through the pipeline, shows level config hints on success
- Add config validation at boot: warn when Slack/Discord is enabled
  but webhook URL is missing
- Fix Filament plugin to respect sentinel.filament.enabled config

+++ src/Metrics/StatsdCollector.php
@@ -31,7 +31,7 @@ class StatsdCollector implements MetricsCollector
     public function __construct()
     {
         $this->host = config('sentinel.metrics.statsd.host', '127.0.0.1');
-        $this->port = config('sentinel.metrics.statsd.port', 8125);
+        $this->port = (int) config('sentinel.metrics.statsd.port', 8125);
         $this->protocol = config('sentinel.metrics.statsd.protocol', 'udp');
         $this->prefix = config('sentinel.metrics.prefix', 'sentinel');
     }
